Better internet connectivity check for UserWishListActivity. Refactor: rename WishlistUtenteActivity to UserWishlistActivity and pojo package to util package
