+++ app/src/main/java/com/peppedalterio/whatsyourwish/ContactsListFragment.java
@@ -2,7 +2,6 @@ package com.peppedalterio.whatsyourwish;
 
 import android.content.Intent;
 import android.database.Cursor;
-import android.os.Build;
 import android.os.Bundle;
 import android.provider.ContactsContract;
 import android.support.annotation.NonNull;
@@ -10,26 +9,20 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.telephony.PhoneNumberUtils;
 import android.text.Editable;
-import android.text.Html;
 import android.text.TextWatcher;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Adapter;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
-import android.widget.ListAdapter;
 import android.widget.ListView;
 
-import com.google.android.gms.common.util.NumberUtils;
-import com.peppedalterio.whatsyourwish.pojo.Contact;
-import com.peppedalterio.whatsyourwish.pojo.WishStrings;
+import com.peppedalterio.whatsyourwish.util.Contact;
+import com.peppedalterio.whatsyourwish.util.WishStrings;
 
-import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 import java.util.Objects;
 
 public class ContactsListFragment extends Fragment {
@@ -165,7 +158,7 @@ public class ContactsListFragment extends Fragment {
      * It shows the activity containing selected contact's wishlist.
      */
     private void mostraListaUtente(Contact contact) {
-        Intent myIntent = new Intent(getActivity(), WishlistUtenteActivity.class);
+        Intent myIntent = new Intent(getActivity(), UserWishlistActivity.class);
         myIntent.putExtra("contact", contact);
         this.startActivity(myIntent);
     }
