+++ app/src/main/java/com/peppedalterio/whatsyourwish/util/InternetConnection.java
@@ -1,4 +1,4 @@
-package com.peppedalterio.whatsyourwish.pojo;
+package com.peppedalterio.whatsyourwish.util;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
