+++ app/src/main/java/com/peppedalterio/whatsyourwish/util/WishStrings.java
@@ -1,4 +1,4 @@
-package com.peppedalterio.whatsyourwish.pojo;
+package com.peppedalterio.whatsyourwish.util;
 
 public class WishStrings {
