+++ app/src/main/res/layout/activity_wishlist_utente.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WishlistUtenteActivity">
+    tools:context=".UserWishlistActivity">
 
     <ListView
         android:id="@+id/userwishlist"
